Add schematic and layout for Rev0
This is designed to fit in a hammond 1551F (1551FBK or 1551FGY) case,
with additional machining. There are two versions:

standard: HEDS-5000 connector mounted through-hole.
-smd: HEDS-5000 connector edge-mounted, surface mount.

While both of these will fit in the Hammond case, the -smd version may
incurr lower machining costs, as it does not involve modifications to
the lid, while the standard version is more sturdy.